Fix error in Casing.Style
There are some perhaps unexpected behavioral differences between the
snake_case variants and camel- and PascalCase, whereas the former
enforces casing on _all_ characters in each token, and the latter
expects each token to be correctly cased, and only changes the casing of
the first character if applicable.

It may be necessary to revise this some time, but it is not part of the
public API anyway.