Add Eclipse m2e lifecycle mapping metadata
The Maven plugin now "attaches" itself automatically to the m2e builder
in Eclipse. Currently, this enables it to automatically add
target/generated-test-sources/record-matchers to the Eclipse project's source
folders, removing the need for separately configuring the
builder-helper-maven-plugin for this for automatic discovery in Eclipse.i

Generating the matcher code must still be done with a Maven build.
Apparently, when executing the Mojo in Eclipse, the classpath scanner
does not find anything, so I am guessing there are something variance
kicking in with the way the URLClassLoader for the scanner is built when
executing within Eclipse.